Move ESDocument to TableESDocument
