+++ databuilder/transformer/elasticsearch_document_transformer.py
@@ -2,7 +2,7 @@ from pyhocon import ConfigTree  # noqa: F401
 from typing import Optional  # noqa: F401
 
 from databuilder.transformer.base_transformer import Transformer
-from databuilder.models.elasticsearch_document import ElasticsearchDocument
+from databuilder.models.table_elasticsearch_document import TableESDocument
 from databuilder.models.neo4j_data import Neo4jDataResult
 
 
@@ -28,20 +28,20 @@ class ElasticsearchDocumentTransformer(Transformer):
         if not isinstance(record, Neo4jDataResult):
             raise Exception("ElasticsearchDocumentTransformer expects record of type 'Neo4jDataResult'!")
 
-        elasticsearch_obj = ElasticsearchDocument(elasticsearch_index=self.elasticsearch_index,
-                                                  elasticsearch_type=self.elasticsearch_type,
-                                                  database=record.database,
-                                                  cluster=record.cluster,
-                                                  schema_name=record.schema_name,
-                                                  table_name=record.table_name,
-                                                  table_key=record.table_key,
-                                                  table_description=record.table_description,
-                                                  table_last_updated_epoch=record.table_last_updated_epoch,
-                                                  column_names=record.column_names,
-                                                  column_descriptions=record.column_descriptions,
-                                                  total_usage=record.total_usage,
-                                                  unique_usage=record.unique_usage,
-                                                  tag_names=record.tag_names)
+        elasticsearch_obj = TableESDocument(elasticsearch_index=self.elasticsearch_index,
+                                            elasticsearch_type=self.elasticsearch_type,
+                                            database=record.database,
+                                            cluster=record.cluster,
+                                            schema_name=record.schema_name,
+                                            table_name=record.table_name,
+                                            table_key=record.table_key,
+                                            table_description=record.table_description,
+                                            table_last_updated_epoch=record.table_last_updated_epoch,
+                                            column_names=record.column_names,
+                                            column_descriptions=record.column_descriptions,
+                                            total_usage=record.total_usage,
+                                            unique_usage=record.unique_usage,
+                                            tag_names=record.tag_names)
         return elasticsearch_obj
 
     def get_scope(self):
