+++ tests/unit/transformer/test_elasticsearch_document_transfer.py
@@ -5,6 +5,7 @@ from pyhocon import ConfigFactory  # noqa: F401
 from databuilder import Scoped
 from databuilder.transformer.elasticsearch_document_transformer import ElasticsearchDocumentTransformer
 from databuilder.models.elasticsearch_document import ElasticsearchDocument
+from databuilder.models.table_elasticsearch_document import TableESDocument
 from databuilder.models.neo4j_data import Neo4jDataResult
 
 
@@ -71,21 +72,21 @@ class TestElasticsearchDocumentTransformer(unittest.TestCase):
 
         result = transformer.transform(data)
 
-        expected = ElasticsearchDocument(elasticsearch_index='test_es_index',
-                                         elasticsearch_type='test_es_type',
-                                         database="test_database",
-                                         cluster="test_cluster",
-                                         schema_name="test_schema_name",
-                                         table_name="test_table_name",
-                                         table_key="test_table_key",
-                                         table_last_updated_epoch=123456789,
-                                         table_description="test_table_description",
-                                         column_names=["test_col1", "test_col2"],
-                                         column_descriptions=["test_col_description1",
-                                                              "test_col_description2"],
-                                         total_usage=10,
-                                         unique_usage=5,
-                                         tag_names=["test_tag1", "test_tag2"])
+        expected = TableESDocument(elasticsearch_index='test_es_index',
+                                   elasticsearch_type='test_es_type',
+                                   database="test_database",
+                                   cluster="test_cluster",
+                                   schema_name="test_schema_name",
+                                   table_name="test_table_name",
+                                   table_key="test_table_key",
+                                   table_last_updated_epoch=123456789,
+                                   table_description="test_table_description",
+                                   column_names=["test_col1", "test_col2"],
+                                   column_descriptions=["test_col_description1",
+                                                        "test_col_description2"],
+                                   total_usage=10,
+                                   unique_usage=5,
+                                   tag_names=["test_tag1", "test_tag2"])
 
         self.assertIsInstance(result, ElasticsearchDocument)
         self.assertDictEqual(vars(result), vars(expected))
